Revert "xiph demux: Kill a warning (comparison betw. signed and unsigned)"
This reverts commit ea5e62b58eb0660153f73b0870540e5175d00820.

This removes all code from Andri Pálsson

+++ modules/demux/xiph.h
@@ -48,7 +48,7 @@ static inline int xiph_SplitHeaders(unsigned packet_size[], void *packet[], unsi
         }
         size += packet_size[i];
     }
-    if (end - current < (int)size)
+    if (end - current < size)
         return VLC_EGENERIC;
     packet_size[count - 1] = end - current - size;
